requirements.txt
alembic
fastapi
sqlalchemy
marshmallow
marshmallow-sqlalchemy
psycopg2-binary
pytest
python-dotenv
uvicorn